fix: disable sw in dev
